#572 Rename `MssqlRemote` to `SqlServerRemote` and manage SQL Server driver in dependency management
