+++ gradle.properties
@@ -46,6 +46,7 @@ mariaDbClientVersion=2.7.4
 log4jVersion=2.7
 javaxServletVersion=3.1.0
 jettyVersion=9.4.44.v20210927
+mssqlDriverVersion=9.4.0.jre11
 
 # Gradle plugin versions
 licenserVersion=0.5.0
