+++ hartshorn-data/src/main/java/org/dockbox/hartshorn/data/hibernate/HibernateJpaRepository.java
@@ -31,7 +31,7 @@ import org.dockbox.hartshorn.data.context.EntityContext;
 import org.dockbox.hartshorn.data.jpa.JpaRepository;
 import org.dockbox.hartshorn.data.remote.DerbyFileRemote;
 import org.dockbox.hartshorn.data.remote.MariaDbRemote;
-import org.dockbox.hartshorn.data.remote.MssqlRemote;
+import org.dockbox.hartshorn.data.remote.SqlServerRemote;
 import org.dockbox.hartshorn.data.remote.MySQLRemote;
 import org.dockbox.hartshorn.data.remote.PersistenceConnection;
 import org.dockbox.hartshorn.data.remote.PostgreSQLRemote;
@@ -114,7 +114,7 @@ public class HibernateJpaRepository<T, ID> implements JpaRepository<T, ID>, Enab
         this.registerDialect(MySQLRemote.class, MySQL8Dialect.class.getCanonicalName());
         this.registerDialect(PostgreSQLRemote.class, PostgreSQL95Dialect.class.getCanonicalName());
         this.registerDialect(MariaDbRemote.class, MariaDB103Dialect.class.getCanonicalName());
-        this.registerDialect(MssqlRemote.class, SQLServer2012Dialect.class.getCanonicalName());
+        this.registerDialect(SqlServerRemote.class, SQLServer2012Dialect.class.getCanonicalName());
     }
 
     @Override
