--- hartshorn-data/src/main/java/org/dockbox/hartshorn/data/remote/MssqlRemote.java
@@ -1,42 +1,0 @@
-/*
- *  Copyright (C) 2020 Guus Lieben
- *
- *  This framework is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as
- *  published by the Free Software Foundation, either version 2.1 of the
- *  License, or (at your option) any later version.
- *
- *  This library is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
- *  the GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
- */
-
-package org.dockbox.hartshorn.data.remote;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public final class MssqlRemote extends JdbcRemote {
-
-    public static final MssqlRemote INSTANCE = new MssqlRemote();
-
-    @Override
-    protected String connectionString(JdbcRemoteConfiguration server) {
-        return "jdbc:%s://%s:%s".formatted(this.type(), server.server(), server.port());
-    }
-
-    @Override
-    protected String type() {
-        return "sqlserver";
-    }
-
-    @Override
-    public String driver() {
-        return "com.microsoft.sqlserver.jdbc.SQLServerDriver";
-    }
-}
